+++ lib/dotsync/actions/pull_action.rb
@@ -11,6 +11,7 @@ module Dotsync
       show_env_vars
       show_mappings_legend
       show_mappings
+      show_differences_legend
       show_differences
       return unless options[:apply]